+++ src/components/input/input.styles.ts
@@ -281,8 +281,9 @@ export default css`
     display: none;
   }
 
-  /* Don't show the firefox clear icon in Firefox */
-  .hide-firefox-clear-icon {
+  /** Hide Firefox's clear button on date and time inputs */
+  .input--is-firefox input[type='date'],
+  .input--is-firefox input[type='time'] {
     clip-path: inset(0 2em 0 0);
   }
 `;